document_analyzer_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
pdf2image>=1.16.0
Pillow>=9.0.0
